requirements.txt
aiohttp==3.9.5
customtkinter==5.2.2
PyMuPDF==1.24.5
Pillow==10.3.0
PyPDF2==3.0.1
reportlab==4.2.0
selenium==4.22.0
seleniumbase==4.27.5
